moved the input and output handlers out of AndroidBlue and into their own classes
