+++ ProjectSpartanV2/.idea/modules.xml
@@ -7,4 +7,3 @@
     </modules>
   </component>
 </project>
-
