+++ ProjectSpartanV2/.idea/vcs.xml
@@ -4,4 +4,3 @@
     <mapping directory="" vcs="" />
   </component>
 </project>
-
