+++ ProjectSpartanV2/app/src/main/java/com/haloproject/bluetooth/AndroidBlue.java
@@ -15,6 +15,11 @@ import android.os.Looper;
 import android.util.Log;
 import android.widget.ArrayAdapter;
 
+import com.haloproject.bluetooth.InputHandlers.BeagleAutoOffSwitch;
+import com.haloproject.bluetooth.InputHandlers.BeagleAutoSwitch;
+import com.haloproject.bluetooth.InputHandlers.BeagleSwitch;
+import com.haloproject.bluetooth.OutputHandlers.BeagleDoubleOutput;
+import com.haloproject.bluetooth.OutputHandlers.BeagleIntegerOutput;
 import com.haloproject.projectspartanv2.SoundMessageHandler;
 import com.haloproject.projectspartanv2.Warning;
 
@@ -22,6 +27,7 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.IOException;
+import java.io.OutputStream;
 import java.lang.reflect.Method;
 import java.util.LinkedList;
 import java.util.List;
@@ -430,139 +436,17 @@ public class AndroidBlue {
             }
         }
     };
-    
-    public class BeagleDoubleOutput {
-        public BeagleDoubleOutput(String location) {
-            this.location = location;
-        }
 
-        private String location;
-
-        public double getValue() {
-            try {
-                return mJSON.getDouble(location); //get double could accept and integer because double is bigger than an integer. thus this code will work for both ints and doubles
-            } catch (Exception e) {
-                return -1000.0;
-            }
-        }
+    /**
+     * returns a reference to the JSON object inside androidBlue.
+     * not save because we are returning a reference.
+     * */
+    public JSONObject getJSON() {
+        //is JSONObject immutable?
+        return mJSON; //TODO: look into returning a copy (how will this effect performance)
     }
 
-    public class BeagleIntegerOutput {
-        public BeagleIntegerOutput(String location) {
-            this.location = location;
-        }
-
-        private String location;
-
-        public int getValue() {
-            try {
-                return mJSON.getInt(location); //get double could accept and integer because double is bigger than an integer. thus this code will work for both ints and doubles
-            } catch (Exception e) {
-                return -1;
-            }
-        }
-    }
-
-    abstract private class BeagleInputHandler {
-        private BeagleInputHandler(String location) {
-            this.location = location;
-        }
-
-        protected String location;
-
-        protected void setState(String state) {
-            try {
-                JSONObject switchObject = new JSONObject();
-                switchObject.put(location, state);
-                mSocket.getOutputStream().write(switchObject.toString().getBytes());
-            } catch (Exception e) {
-
-            }
-        }
-
-        protected boolean isStateSet(String state) {
-            try {
-                return mJSON.getString(location).equals(state);
-            } catch (Exception e) {
-                return false;
-            }
-        }
-    }
-
-    //turn switched auto or off
-    public class BeagleAutoOffSwitch extends BeagleInputHandler {
-        public BeagleAutoOffSwitch(String location) {
-            super(location);
-        }
-
-        public void auto() {
-            setState("auto");
-        }
-
-        public void off() {
-            setState("off");
-        }
-
-        public boolean isAuto() {
-            return isStateSet("auto");
-        }
-
-        public boolean isOff() {
-            return isStateSet("off");
-        }
-    }
-
-    //turns things on or off or auto
-    public class BeagleAutoSwitch extends BeagleInputHandler {
-        public BeagleAutoSwitch(String location) {
-            super(location);
-        }
-
-        public void auto() {
-            setState("auto");
-        }
-
-        public void on() {
-            setState("on");
-        }
-
-        public void off() {
-            setState("off");
-        }
-
-        public boolean isAuto() {
-            return isStateSet("auto");
-        }
-
-        public boolean isOn() {
-            return isStateSet("on");
-        }
-
-        public boolean isOff() {
-            return isStateSet("off");
-        }
-    }
-
-    //used for turning things on or off on the beaglebone
-    public class BeagleSwitch extends BeagleInputHandler {
-        public BeagleSwitch(String location) {
-            super(location);
-        }
-
-        public void on() {
-            setState("on");
-        }
-
-        public void off() {
-            setState("off");
-        }
-
-        public boolean isOn() {
-            return isStateSet("on");
-        }
-
-        public boolean isOff() {
-            return isStateSet("off");
-        }
+    public OutputStream getOutputStream() throws IOException {
+        return mSocket.getOutputStream();
     }
 }
